Make children optional, correctness changes
The children can also be any ReactNode in React 16, as opposed to only JSX.Element. Even in React 15, the current definitions exclude `string` and `number`.

`children` is optional, especially when `render` is provided.

The `onChange` is passed through to the top level element so if the `tag` happens to be one with an `onChange` event, the `onChange` callback may be called with `React.FormEvent`.