feat: export .mts for test-utils
Compile `test-utils` with tsup instead of plain `tsc`, and add the `test-utils.d.mts` file to the types in package.json. This should ensure it can be consumed correctly in a modern setup. Keeps the old adjacent `test-utils.d.ts` for legacy.

The `tsconfig.test.json` is no longer needed